snapper: Add support for specifying snapper config name.
Implement a new "snapper_config" setting for specifying which config
the plugin uses for creating snapshots. This allows the plugin to
function on systems where the root snapper config is named something
other than "root".

The setting defaults to "root" (as before) when not set.

= changelog =
msg:           snapper: Add support for snapper configs not named "root".
type:          bugfix
resolves: